Move configurePersistence code to blacklist module
- exported from entrypoint as named export 'configurePersistence'
- each module can add an optional 'persistence.js' file containing the redux store->local storage/pouchDB persistence logic
