+++ src/options/blacklist/index.js
@@ -1,5 +1,5 @@
 import * as actions from './actions'
 export { default as reducer } from './reducers'
+export { default as configurePersistence } from './persistence'
 export { default } from './container'
 export { actions }
-
